+++ src/components/VideoPlayer.js
@@ -6,6 +6,9 @@ EVENT TO WATCH
 play, pause, seeking, seeked, volumeChange Waiting
 
 onBUFFERING 
+
+audio has a durationchange Event
+todo: find a way to figure out if its a direct load then mute the audio.
 */
 
 const VideoPlayer = ({ videoUrl, audioUrl, poster }) => {
@@ -14,58 +17,48 @@ const VideoPlayer = ({ videoUrl, audioUrl, poster }) => {
 	const videoPlayer = useRef();
 	const audioPlayer = useRef();
 
-	const play = useCallback(() => {
+	const play = () => {
 		videoPlayer.current.play().catch(console.log);
 		audioPlayer.current.play().catch(console.log);
-		// console.log({ paused: videoPlayer.paused });
-		return Object.is(videoPlayer.paused, undefined) ? -1 : 0;
-	}, []);
+	};
 
 	const pause = () => {
 		videoPlayer.current.pause();
 		audioPlayer.current.pause();
 	};
 
-	// autoplay implmentation.
-	useEffect(() => {
-		if (!autoPlay) return null;
-		audioPlayer.current.addEventListener("canplay", () => {
-			videoPlayer.current.addEventListener("canplay", () => {
-				// console.log(play());
-				if (play() === -1) {
-					// did not play most likely because of -> DOMException: play() failed because the user didn't interact with the document first
-					// we need the muted property to play without user interaction
-					// doesnt work because the audio doesnt play automatically.
-					// this is good but lets just remove this autoplay tag can take care.
-					videoPlayer.current.muted = true;
-					play();
-				}
-			});
-		});
-	}, [autoPlay, play]);
+	const deb = (e) => {
+		console.log(e);
+	};
+
+	const changeTimeStamp = (e) => {
+		let t = e.target.currentTime;
+		audioPlayer.current.currentTime = t;
+	};
 
 	return (
 		<>
 			<video
+				height="400px"
+				width="600px"
+				ref={videoPlayer}
 				className="video-player"
+				autoPlay={autoPlay}
 				poster={poster}
 				preload={shouldPreLoad}
-				ref={videoPlayer}
 				loop
+				controls
 				onPlay={play}
 				onPause={pause}
-				controls
-				autoPlay={autoPlay}
+				onWaiting={pause}
+				onCanPlayThrough={play} // in response to onWaiting
+				onStalled={pause}
+				onSeeked={changeTimeStamp}
 			>
 				<source src={videoUrl} type="" />
 				your browser does not support video
 			</video>
-			<audio
-				src={audioUrl}
-				ref={audioPlayer}
-				// style={{ display: "none" }}
-				loop
-			></audio>
+			<audio src={audioUrl} ref={audioPlayer} loop></audio>
 		</>
 	);
 };